+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Enable MegaLinter LLM Advisor
+- Org Manager: Add contextual menu to set alias
 
 ## [6.6.1] 2024-09-20